added in the main part of my program
This is what I wanted - a (very) simple ff game-book kind of program which has some choice and uses randomly generated monsters which differ according to where they are encountered.

I'll come back to it when I've learnt how to put the fight method into the Character class. The problem at the moment is I'm not sure how to pass values from sub-classes to methods in their super-classes.